backend/requirements.txt
Django>=4.0,<5.0
gunicorn>=20.0.4
daphne>=4.1.2
mysqlclient>=2.0
django-allauth>=0.51
requests>=2.32.0
PyJWT>=2.9.0
cryptography>=41.0.0
djangorestframework>=3.15.0
django-filter>=24.0
django-cors-headers>=4.0
dj-rest-auth>=5.0.0
djangorestframework-simplejwt>=5.3.0
django-environ>=0.11.2
psycopg2-binary>=2.9
pillow>=11.2.1
